+++ render_all.jsx
@@ -1,4 +1,5 @@
-#include "scripts/render.jsx";
+#include "scripts/jsx/render.jsx";
+#include "settings.jsx";
 
 // File path to main working directory
 var file_path = File($.fileName).parent.fsName;