+++ scripts/jsx/render.jsx
@@ -1,9 +1,8 @@
 // Instantiate the template map
 var class_template_map = {};
 
-#include "json2.js";
-#include "../settings.jsx";
-#include "layouts.jsx";
+#include "scripts/jsx/json2.js";
+#include "scripts/jsx/layouts.jsx";
 
 function retrieve_card_name_and_artist(file) {
     /**
@@ -52,10 +51,10 @@ function call_python(card_name, file_path, set) {
      */
 
     // default to Windows command
-    var python_command = "py \"" + file_path + "/scripts/get_card_info.py\" \"" + card_name + "\" \"" + set + "\"";
+    var python_command = python_exe + " \"" + file_path + "\\py\\get_card_info.py\" \"" + card_name + " set:" + set + "\" " + output_directory + "card.json";
     if ($.os.search(/windows/i) === -1) {
         // macOS
-        python_command = "/usr/local/bin/python3 \"" + file_path + "/scripts/get_card_info.py\" \"" + card_name + "\" \"" + set + "\" >> " + file_path + "/scripts/debug.log 2>&1";
+        python_command = "/usr/local/bin/python3 \"" + file_path + "/py/get_card_info.py\" \"" + card_name + "\" \"" + set + "\" >> " + file_path + "/py/debug.log 2>&1";
     }
     app.system(python_command);
 
@@ -70,7 +69,10 @@ function call_python(card_name, file_path, set) {
             "Try running the command from the command line as that may help you debug the issue"
         );
     }
-    return JSON.parse(JSON.parse(json_string));
+
+    // TODO: Match JSON card name to card_name as a verification step
+
+    return JSON.parse(json_string);
 }
 
 function call_python_set(card_set, file_path) {
@@ -79,10 +81,10 @@ function call_python_set(card_set, file_path) {
      */
 
     // default to Windows command
-    var python_command = "py \"" + file_path + "/scripts/get_set_info.py\" \"" + card_set + "\"";
+    var python_command = python_exe + " \"" + file_path + "/py/get_set_info.py\" \"" + card_set + "\" " + output_directory + "set.json";;
     if ($.os.search(/windows/i) === -1) {
         // macOS
-        python_command = "/usr/local/bin/python3 \"" + file_path + "/scripts/get_set_info.py\" \"" + card_set + "\" >> " + file_path + "/scripts/debug.log 2>&1";
+        python_command = "/usr/local/bin/python3 \"" + file_path + "/py/get_set_info.py\" \"" + card_set + "\" >> " + file_path + "/py/debug.log 2>&1";
     }
     app.system(python_command);
 
@@ -92,12 +94,12 @@ function call_python_set(card_set, file_path) {
     json_file.close();
     if (json_string === "") {
         throw new Error(
-            "\n\ncard.json does not exist - the system failed to successfully run get_card_info.py.\nThe attempted Python call was made with the " +
+            "\n\nset.json does not exist - the system failed to successfully run get_card_info.py.\nThe attempted Python call was made with the " +
             "following command:\n\n" + python_command + "\n\nYou may need to edit this command in render.jsx depending on your computer's configuration. " +
             "Try running the command from the command line as that may help you debug the issue"
         );
     }
-    return JSON.parse(JSON.parse(json_string));
+    return JSON.parse(json_string);
 }
 
 function select_template(layout, file, file_path, new_template) {
@@ -144,7 +146,7 @@ function render(file,current_template) {
         if (layout_name in layout_map) {
             var layout = new layout_map[layout_name](scryfall, card_name);
         } else {
-            throw new Error("Layout" + layout_name + " is not supported. Sorry!");
+            throw new Error("Layout " + layout_name + " is not supported. Sorry!");
         }
 
         // if artist specified in file name, insert the specified artist into layout obj